// src/app/models/Issue.ts
export class Issue {
  id?: string;
  title?: string;
  description?: string;
  reported_at?: string;
  reporter_id?: string;
  created_at?: string;
  criticality?: string;
  verified_by?: string;
  target_date?: string;
  closure_date?: string;
  status?: string;
  pending_with?: string;
  system?: string;
}